+++ src/Commands/Fun/Say.ts
@@ -1,6 +1,6 @@
 import Command from '../../Structures/Command';
 import Embed from '../../Structures/Embed';
-import { Message, MessageEmbed } from 'discord.js';
+import { Message } from 'discord.js';
 
 export default class extends Command {
     constructor() {
@@ -12,19 +12,19 @@ export default class extends Command {
         );
     }
 
-    init(message: Message, args: string[]): Promise<Message> {
+    init(message: Message, args: string[]) {
         if(!super.hasPermissions(message)) {
             return message.channel.send(Embed.missing_perms(this.permissions))
         } else if(args.length < 1) {
             return message.channel.send(Embed.missing_args(1, this.name, [
                 'Hello!', 'Goodbye!'
-            ]))
+            ]));
         }
 
-        return message.channel.send(this.formatEmbed(message, args))
+        return message.channel.send(this.formatEmbed(message, args));
     }
 
-    formatEmbed(message: Message, args: string[]): MessageEmbed {
+    formatEmbed(message: Message, args: string[]) {
         const embed = Embed.success()
             .setTimestamp()
             .setDescription(`