+++ src/Commands/Server/UserInfo.ts
@@ -1,5 +1,5 @@
 import Command from '../../Structures/Command';
-import { Message, MessageEmbed, GuildMember, Activity } from 'discord.js';
+import { Message, GuildMember, Activity } from 'discord.js';
 import Embed from '../../Structures/Embed';
 
 export default class extends Command {
@@ -12,7 +12,7 @@ export default class extends Command {
         );
     }
 
-    async init(message: Message, args: string[]): Promise<Message> {
+    async init(message: Message, args: string[]) {
         if(!super.hasPermissions(message)) {
             return message.channel.send(Embed.missing_perms(this.permissions));
         }
@@ -20,7 +20,7 @@ export default class extends Command {
         return message.channel.send(await this.formatEmbed(message, args.shift()));
     }
 
-    async formatEmbed(message: Message, id: string): Promise<MessageEmbed> {
+    async formatEmbed(message: Message, id: string) {
         let member: GuildMember; // = message.mentions.members?.first() ?? message.member;
         if(id && !message.mentions.members.first()) {
             try {
@@ -42,18 +42,18 @@ export default class extends Command {
             \`\`${perms}\`\`
             `)
             .setThumbnail(member.user.displayAvatarURL() ?? member.user.defaultAvatarURL)
-            .addField('**Tag:**', member.user.tag, true)
+            .addField('**Tag:**',        member.user.tag, true)
             .addField('**Role Color:**', member.displayHexColor, true)
-            .addField('**ID:**', member.id, true)
-            .addField('**Discrim:**', member.user.discriminator, true)
-            .addField('**Nickname:**', member.nickname ?? 'None', true)
-            .addField('**Bot:**', member.user.bot ? 'Yes' : 'No', true)
+            .addField('**ID:**',         member.id, true)
+            .addField('**Discrim:**',    member.user.discriminator, true)
+            .addField('**Nickname:**',   member.nickname ?? 'None', true)
+            .addField('**Bot:**',        member.user.bot ? 'Yes' : 'No', true)
         
         return embed;
     }
 
-    formatPresence(activities: Activity[]): string {
-        const push = [];
+    formatPresence(activities: Activity[]) {
+        const push: string[] = [];
         for(const activity of activities) {
             switch(activity.type) {
                 case 'CUSTOM_STATUS':