+++ src/Commands/Settings/Create.ts
@@ -1,6 +1,6 @@
 import Command from '../../Structures/Command';
-import { dbHelpers } from '../../Structures/GuildSettings/GuildSettings';
-import { Message, MessageEmbed } from 'discord.js';
+import { dbHelpers } from '../../Helpers/GuildSettings';
+import { Message } from 'discord.js';
 import Embed from '../../Structures/Embed';
 
 export default class extends Command {
@@ -13,7 +13,7 @@ export default class extends Command {
         );
     }
 
-    async init(message: Message): Promise<any> {
+    async init(message: Message) {
         if((!super.hasPermissions(message) || !super.userHasPerms(message, [ 'ADMINISTRATOR' ]))
             && !this.isBotOwner(message.author.id)
         ) {
@@ -24,17 +24,11 @@ export default class extends Command {
 
         const value = dbHelpers.set(message);
         if(value.changes === 1) {
-            return message.channel.send(this.formatEmbed());
+            return message.channel.send(Embed.success('Guild settings are now available for you to use!'));
         } else if(value.changes === 0) {
             return message.channel.send(Embed.fail('Guild has already been inserted into database!'));
         }
 
         return message.channel.send(Embed.fail('An unknown error occurred!'))
-    }
-
-    formatEmbed(): MessageEmbed {
-        const embed = Embed.success('Guild settings are now available for you to use!');
-
-        return embed;
     }
 }