+++ src/Helpers/load.env.ts
@@ -1,4 +1,3 @@
-import KhafraError from '../Structures/Error';
 import { readFileSync, existsSync } from 'fs';
 import { join } from 'path';
 import { EOL } from 'os';
@@ -6,7 +5,7 @@ import { EOL } from 'os';
 export default function loadEnv() {
     const path = join(process.cwd(), '.env');
     if(!existsSync(path)) {
-        new KhafraError('.env', 'No .env file found at the root of the repo!', 'load.env.js');
+       throw new Error('.env: No .env file found at the root of the repo!');
     }
 
     const buffer = readFileSync(path).toString().split(EOL);