+++ src/Structures/Database.ts
@@ -1,6 +1,6 @@
 import { join } from 'path';
 import betterSQLite3 = require('better-sqlite3');
 
-const db = betterSQLite3(join(process.cwd(), 'build/Structures/GuildSettings/guildsettings.db'));
+const db = betterSQLite3(join(process.cwd(), 'build/Structures/guildsettings.db'));
 
 export default db;