+++ CHANGELOG.md
@@ -11,3 +11,9 @@
 * Build scripts are now available for ease of access.
 * Intelligently caches GuildSettings.
 * Lots of logic and parsing errors were fixed.
+
+# v1.0.0~beta2
+* Cleanup on some commands and functions.
+* GuildSettings: react to a message for a role. New message, old message, guild emoji, unicode emoji. Tweak this setting to your liking.
+* Include typings for twemoji-parser.
+* Remove inferred return types, add in other types.